roommanager: Add setting to prefer the local node for selection
It may be desired to prefer using the local node when creating a new
room; add a setting that uses the current signaling node when creating a
new room. The availability of the node is still confirmed before
selecting it.

If using the system load selector, a preferred node will not be selected
if it does not fall below the configured load threshold.

This helps with scenarious where the LiveKit server is being selected
with a regional load balancer to ensure a more distant node isn't
randomly chosen.

Includes tests for the system load selector.