Refactoring connection stats
Had repeated code and some bugs. So consolidating and fixing what I
could find. Hopefully in the process simplifying. This increases
the RTCP packet processing a bit as the packets are run through twice,
but RTCP is rare.

Still finding one issue on migration where the participant who is
not migrating drops in quality for one cycle and then recovers. But,
these changes seem to have fixed the periodic quality oscillation after
migration (as far as I could test) which is good.

The one time drop is to due a huge packet loss. Probably something when
switching to a different remote media track because of not resetting.
Will look at that in a different PR.

Hopefully, this should catch the correct upload bytes too.