Need to return delta the new provisional layer takes.
As provisional allocation runs through layers for all down tracks,
as layers go up, if a down track is able to use a higher layer,
it needs to return how much additional bandwidth it is taking up.
Not the full value as the stream allocator has already allocated
for a down track at a previously allocated layer.