Coalesce track events into stream allocator.
In a large room, when a participant joins, that participant could
subscribe to a significant number of tracks in a burst upon join.
This could generate a bunch of events into stream allocator.

Coalesce events of a track so that only one allocate event per track
is posted to channel, i. e. prevent posting more allocate events
till the previous one is processed. As allocate will take view
of current state of track when processing the event, there is no
need to post multiple events.