Do not re-use transceiver when negotiation is pending.
When a track is unpublished and published again, it is
possible that track events are not fired on client side
due to renegotiation debounce and transceiver re-use.
Although SSRC should change, looks like a re-use of
transceiver is not firing events.

So, when a renegotiation is pending, do not re-use
transceiver. This does affect efficiency of SDP when
there are too many track subscribe/unsubsribe is bunched
in a debounce interval. But, hopefully it is not a big
issue in real life scenarios. One option is to reduce
the debounce interval (from the current 150 ms) a little
bit to trade off potentially more re-negotiations for SDP
efficiency.