Check participant SID also while removing a participant.
It is possible that leave request from a different participant
session closes an existing participant session even if the
participant SID is not matching. Prevent that by checking
for participant SID match before removing.

NOTE: RoomService path is intended to remove by identity.
So, it will remove participant if present by identity irrespective
of SID.