Track subscription operations per source track.
Subscriptions are a bit complex. There are a few challenges to tackle
- Repeated subscriptions
- Switching subscription from remote -> local or vice-versa
- Multiple unsubscriptions (because of removing subscriber and removing
  all subscribers when publishing track goes away)
- As all subscribe operations are queued, cannot determine if
  something is a duplicate when queuing the operation.

Given these, doing a simple desired vs actual becomes quite tricky.
Actually, that is one of the reasons I had done the transitions queue
and ensuring expected transitions happen accounting for duplicate
transitions too.

From logs of sessions, the only issue seems to be unsubscribe timing
out. That happens with migration when source switches from remote ->
local or vice-versa because the previous source does an unsubscribe
transisiton, but the new source is already set.

The solution to that is to maintain expected transitions per source
track and ensure those happen properly. Implementing that change here.

Also a couple of other things
- Some clean up
- Introduce a config to issue a full reconnect on subscription failure.
  Not enabled by default. Can be enabled after monitoring sessions for
  false negatives.