Minor tweaks to frame based stream tracker.
- Do status check if the interval is more than 98% of eval interval.
  If the eval interval is set close to check interval, things get off
  by a few ms and it misses a check window. Add some tolerance.

Could not find a good way to use inter-layer fps to eliminate any config.
Things change randomly (at least under NLC testing) which makes it hard to
rely on inter-layer. And browser variations like some browsers stopping lower
layers makes it error prone.

The single config of `MinFPS` with sane defaults seems like it could be the way.
Hopefully, the default will be good for an overwhelming majoirty of use cases.
If not, we can look at adding one config to `VideoLayers` in `TrackInfo` for `MinFPS`
and override the server side config per track. That is a small enough change to the
protocol and fairly straight forward to document/understand.