Account for time before measurement available in connection quality.
There is a window of time where a layer is declared available before
bitrate measurement is available. That window could be 1 - 1.5 seconds
which could lead connection quality to drop because of expected
distance.

Account for it by doing two things
- Looking for available layers.
- Notifying on available layer change event.

This probably needs to be done for down track side also. That will be
a bigger change. Will look at it in a different PR.