Unify the forwarder between dependency descriptor and no DD case.
Had the two paths different while moving to opportunistic forwarding
for a couple of reasons (did not want to affect the DD path and did not
want to get slowed down by not a fully released feature).

Unifying the paths now. But, this is not quite working well for DD case.

Seeing a few things that needs investigation/more work
- Receive stream start up takes longer - looks like forwarding starts
  early but decoder is not able to decode/render. Unclear why.
- The decode side goes through several resolution steps (spaced about
  10 seconds apart). With JS sample app default, the receive side should
  end up with a adaptive stream resolution which should pick layer 1.
  But, before adaptive stream kicks in, server starts with LOW, then
  switches to HIGH and then eventually comes to layer 1. Subscribed
  resolutions go from 160x90 -> 240x135 -> one more in between ->
  640x360 which is the target for layer 1.
- When it finally settles at 640x360 on subscriber side, the publisher
  side bit rate shows 1.8 Mbps. JS sample app console logs say that
  layer 2 is disabled which is the correct dynacast behaviour. Unclear
  what is driving the bitrate on publisher side to be that high.

Also swithing receiver side to consistently use layer 0(buffer0) for
SVC.

Testing:
--------
Sanity check that video is received for camera/screen share when using
AV1 or VP8 or H264. (with AV1 issues above which need further
investigation).